fix(zsh): change plugins
shell was stuttering on very long or very short
commands. i felt like it was on completion,
because other terminal apps worked flawlessly
(not kitty problem) and other shell (bash) was fine.

i thought that might be powerlevel10k (because it
is not being actively maintained anymore), and
changed the prompt to oh-my-posh. stuttering
still persisted.

and after changing zsh-autocomplete to
zsh-completions with this snippet, everything
works normal again now.