feat(hypr): change default swapping to silent
"shift" stands for "shift the window to", and "shift+ctrl" stands for
"shift the window to and control my position". this acts as a modifier
for the action alongside the swap.

also, special workspace bind was incorrect. and this way this is more
comfortable, because for the special workspace you more often will need
just to send the window and forget about it, and that's one less button
to press.