feat(kitty): remove tab bar and window borders
very minimalistic, looks clean and gives more space.

cons:

- active window is not obvious
- because of no borders, the programs can be not so obvious either

if second con is a deal breaker, window_border_width can be set to 1px.

also, no tab bar? how to see all tabs? it will be fixed in the following
commit, but it's just a bind to show tabs and change to one of them.