refactor(nvim): plugins restructure
first priority of this change is performance:

- lazy load everything if possible
- minimize the number of lines and files before the lazy loading

i seen this approach in nvchad, and wanted to give it a shot. result is
positive: config loads much faster.

minimal number of files minimizes the number of syscalls, which are very
expensive. number of lines is not so important, but overall better
anyway.