feat!: set nushell as SHELL
otherwise kitty opens zsh, and there is no convenient way in niri to
launch kitty with other envs (aside from "sh -c ...").

also it caused the terminal to use zsh when opening directory from a browser,
or just anywhere in the system where the terminal is not launched
manually.

it may cause issues in some apps (but the specification does not say
that SHELL should be a POSIX-compatible shell), but still