fix: don't use uwsm
for some reason, after the niri update uwsm broke for me.

i have shell variable set to /usr/bin/nu, and it is set to this value if
i do printenv, but the shell in the terminal is zsh. how?

and if i set shell to nushell in kitty manually, then env from zsh is
not sourced. either duplicate the environment (and maybe that's a better
solution) for each shell or just don't use uwsm for now.

if i don't get "waiting for user session" in a while, i consider that
uwsm is not needed for me. uuctl is cool and overall shutdown sequence
as well, but i don't want to configure the shell right now.